python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telemachus-py"
version = "0.1.0"
description = "Telemachus: an open telematics pivot format (manifest YAML + Parquet tables) with Python tools to export, validate and read datasets."
authors = [{name = "Telemachus Project"}]
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.7",
    "pandas>=2.1",
    "pyarrow>=16.0",
    "click>=8.1",
    "pyyaml>=6.0",
    "jsonschema>=4.22"
]

[project.scripts]
tele = "telemachus.cli:tele"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "ipykernel"
]
viz = [
    "matplotlib"
]
geo = [
    "geopandas",
    "shapely",
    "pyproj"
]

[tool.setuptools.packages.find]
include = ["telemachus*"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]